move entry point to __main__ file
